+++ click_extended/interactive/__init__.py
@@ -1,0 +1,5 @@
+"""Initialization file for the 'click_extended.interactive' module."""
+
+from click_extended.interactive.selection import selection
+
+__all__ = ["selection"]